Partially Completed or Failed Projects/Speakers Detection from audio or video/requirements.txt
numpy>=1.20.0
matplotlib>=3.5.0
openai-whisper>=20231117
torch>=2.0.0
librosa>=0.9.0
soundfile>=0.12.0
pandas>=1.5.0
seaborn>=0.12.0
pyannote.audio>=2.1.1
scikit-learn>=1.0.0
moviepy>=1.0.3